crop_disease_backend/requirements.txt
# requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
requests>=2.28.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-flask>=1.2.0
motor>=3.0.0
pymongo>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0